Use `lsp_types` directly in `sway-lsp` tests instead of simulating a client with `tower-lsp` (#4976)
## Description
This PR removes the mock client <> server interaction to test the
language server features in favour of using `lsp_types` directly. There
are a few places like `did_change` where we are still using the
exisiting framework, this is because we still rely on the `tower-lsp`
event loop to synchronise notification messages. This will be tackled
with the custom event loop work. Should be easy to just update the
remaining tests here within that PR. I've also kept the publish
diagnostics test the same as we still rely on the `Client` from
`tower-lsp` for that functionality.

There were also a few tests here that were unnecessary. Namely,
`initialized` `initializes_only_once` `shutdown`
`refuses_requests_after_shutdown`. These tests are actually duplicated
from `tower-lsp` itself, so there is really no reason we need to run
them again in `sway-lsp`.
 
working towards #4958