Parallelize traversing AST's in the language server (#5001)
## Description
This PR introduces using rayon's parallel iterators when traversing the
lexed and parsed AST's. I initially extended this to also occur on the
typed AST but was running into deadlock issues with how DashMap works.

closes #4959

| Request Type/Method    | Previous  | New      | Improvement % |
|------------------------|-----------|----------|---------------|
| compile_and_traverse   | 2.7732s   | 2.5632s  | 7.57%         |

**edit**

I've split the compilation and traversal stages so they can be
benchmarked independently. Here are the actual traversal performance
results.

| Request Type/Method    | Previous  | New      | Improvement % |
|------------------------|-----------|----------|---------------|
| traverse   | 422ms   | 211ms  | 50.0%         |