Fixes generic types unification of empty structs.
Structs generic type were not unified with type annotation this caused the
struct generic types not beeing resolved to the previous type annotations.

Example:
```let _empty_hash: GenericEmptyStruct<u64> = GenericEmptyStruct {};```

The example above would throw the error `Trait "Hash" is not implemented for type "T".`
While the compiler should know the generic type T is in fact u64.

The solution was to unify the struct type id with the type annotation.
This is now working for variable with type ascriptions, function parameters
, struct fields instantiation and enum instantiation.