Fix rename for alias and function prefixes (#5118)
## Description

Closes https://github.com/FuelLabs/sway/issues/4982

There were actually two issues here. 
1. Type aliases were not getting renamed
2. In any function call with a prefix, the prefix was not getting
renamed, whether it was a type alias or not

I added a test that covers both. I also noticed that the other rename
tests do not assert anything about the response. I think we should at
least assert the number of changes is correct to prevent regressions. I
can add that to this PR or open a follow up.

This changes the behavior of go to the definition of a function call
prefix. Previously it would go to the impl of the prefix where the
function was implemented. Now it takes you to the definition of the
struct/enum, or the top of the module instead. The new behavior is more
similar to rust-analyzer.

## Checklist

- [x] I have linked to any relevant issues.
- [x] I have commented my code, particularly in hard-to-understand
areas.
- [ ] I have updated the documentation where relevant (API docs, the
reference, and the Sway book).
- [x] I have added tests that prove my fix is effective or that my
feature works.
- [ ] I have added (or requested a maintainer to add) the necessary
`Breaking*` or `New Feature` labels where relevant.
- [x] I have done my best to ensure that my PR adheres to [the Fuel Labs
Code Review
Standards](https://github.com/FuelLabs/rfcs/blob/master/text/code-standards/external-contributors.md).
- [x] I have requested a review from the relevant team or maintainers.