Do not put AsmBlocks in the arena
AsmBlocks are part of an Instruction. When an Instruction
is cloned, they must be too. But if they're in the Arena,
when an instruction is cloned, we end up with the same AsmBlock
being part of two instructions. If a transformation modifies
one of it, the other gets modified too, which is incorrect.