Fixes missing error on type unification.
While using generic blanket implementation no error was thrown when another
type other than the generic blanket expected type was used.

This was caused by the typechecker being too permissive in UnknownGeneric
unification. The fix changes the unifier to handle unification with the expected
self and generics as special cases.

Closes #5212