Refactor match expressions and fix critical issue  (#5213)
## Description

This PR fixes the critical issue #5122. It refactors the initial type
checking step in the `matcher` by capturing the variable declarations
and requirements in a tree structure that preserves the relationship
between the requirements and declarations. This structure is traversed
later on in other parts of type checking and desugaring. The desugaring
is also fully refactored and produces semantically correct `if-else`
chain in case of OR patterns.

The high level overview of the new type checking and desugaring together
with examples of desugared `match` expressions can be found in the
_match_expression/typed/mod.rs_ file.

Closes #5122.

## Checklist

- [x] I have linked to any relevant issues.
- [x] I have commented my code, particularly in hard-to-understand
areas.
- [ ] I have updated the documentation where relevant (API docs, the
reference, and the Sway book).
- [x] I have added tests that prove my fix is effective or that my
feature works.
- [ ] I have added (or requested a maintainer to add) the necessary
`Breaking*` or `New Feature` labels where relevant.
- [x] I have done my best to ensure that my PR adheres to [the Fuel Labs
Code Review
Standards](https://github.com/FuelLabs/rfcs/blob/master/text/code-standards/external-contributors.md).
- [x] I have requested a review from the relevant team or maintainers.