Initialise large enough, same-const-initialised, arrays using loops (#5240)
## Description
Closes #5229 

Previously, these arrays were initialised with repeated "load index from
data area, store initialise constant to array". Now we just use a loop
to store the constant initialise to every element of the array. To avoid
overheads of the loop, it's only done, currently, when the array size is
greater than 5 (can be tuned later).