Fixes blanket trait conflicts with impl trait for struct.
When the same trait was implemented for a blanket implementation and for
a struct a conflict occured.

With this commit we use `UnifyCheck::non_generic_constraint_subset` instead of
`UnifyCheck::constraint_subset` this no longer unifies the generic type from the
blanket impl with the struct impl and the conflict no longer occurs.

Closes #5244