correctly deal with missing semicolons (#5236)
## Description

This PR closes https://github.com/FuelLabs/sway/issues/5184, but I have
the impression the same problem can happen with const in other places,
and maybe even with others items.

The problem is that the parsing of `;` is very inconsistent at the
moment. For example, `ItemConst` parses the semicolon, but only as a
`peek`. This because a "parent parser" is actually deciding if the
semicolon is needed or not.

That would be nice if it were more consistent.

## Checklist

- [x] I have linked to any relevant issues.
- [x] I have commented my code, particularly in hard-to-understand
areas.
- [x] I have updated the documentation where relevant (API docs, the
reference, and the Sway book).
- [x] I have added tests that prove my fix is effective or that my
feature works.
- [x] I have added (or requested a maintainer to add) the necessary
`Breaking*` or `New Feature` labels where relevant.
- [x] I have done my best to ensure that my PR adheres to [the Fuel Labs
Code Review
Standards](https://github.com/FuelLabs/rfcs/blob/master/text/code-standards/external-contributors.md).
- [x] I have requested a review from the relevant team or maintainers.

---------

Co-authored-by: Joshua Batty <joshpbatty@gmail.com>