Optimizes `type_engine.insert` to be done inside if.
`check_if_trait_constraints_are_satisfied_for_type` was inserting a `TypeInfo::Custom`
for every trait impl every time it was called.

This change moves the insertion to be done after the unify check has passed.

This optimization reduces the number of size of the type engine from
205665 to 138725 while compiling `/should_pass/language/supertraits` which uses std lib.
The overall improvement is 32.5% less size.