Add AST module caching support for LSP (#5251)
## Description
This PR supersedes #4988 and builds on backend work done by @tritao in
#4733.

A new `ModuleId` is created for each compiled module. The `SourceId`
type now also stores the `ModuleId` that was used to create it. This
allows us to perform garbage collection on types in the Declaration and
Type engines that refer the provided module.

```rust
if let Some(module_id) = engines.se().get_module_id(&path) {
    engines.clear_module(&module_id);
}
```

This method isn't super cheap to call. My tests had this averaging at
54ms. We are currently performing garbage collection on every key stroke
in LSP. We should look to have another metric such as only clearing once
every 20 keystrokes or so in the future.

I ran benchmarking on recompiling the `doc_comments` example 20 times to
simulate rapid keystrokes during coding. Here are the results.

| Without Caching | With Caching |
|---------------------|------------------|
| 89.489 s             | 4.6236  s         |

A speed up of 1835.49% is certainly great, yet 4.6 seconds still sounds
like an eternity. Still have lots of room to improve but this should
improve our current system dramatically.

closes #4994

Co-authored-by: Joao Matos <joao@tritao.eu>