remove local `fixtures-std` and `fixtures-core` dirs and use `std v0.47.0` for lsp tests (#5274)
## Description
Recently in #5203 we made a local copy of the `std` and `core` library
in the `sway-lsp/tests/fixtures` folder. The motivation was to avoid
needing to update a lot of LSP tests each time there were changes to
these folders.

However, @segfault-magnet and @SilentCicero noticed a bug where the
`forc_pkg::find_within` function was sometimes using these copies over
the official versions which was causing build errors as the package
names were identical.

This PR removes the local copies of these libraries and instead uses the
`std` and `core` libraries from github with the `v0.47.0` tag. This
should resolve the issue and also allow us an easier path forward to
upgrading the LSP tests to target new versions of `core` and `std` in
the future when needed.

```toml
std = { git = "https://github.com/FuelLabs/sway", tag = "v0.47.0" }
``` 

As a side effect, running the LSP tests has now gone from taking ~210
seconds to complete to ~95 seconds on my machine.