Provide useful error for "Cannot find trait Hash in this scope" (#5293)
## Description

This PR closes #5129 by providing a detailed guidelines to programmers
on how to fix the "Cannot find trait XYZ in this scope". The issue is
very prominent in the case of the `Hash` trait as demonstrated in the
#5084.

We need to guide programmers at the moment because fixing the root cause
will take some time. (Changes need to fix the root cause are sketched
[in this
comment](https://github.com/FuelLabs/sway/issues/5129#issuecomment-1820721017)).

The PR:
- replaces the current two misleading and redundant error messages with
a clear instruction how to fix the issue at the call site.
- provides the guidelines in a general case, and not only for the
prominent issue with the `Hash` trait.
- removes the redundant `SymbolNotFound` duplicated error and leaves
only the specific `TraitNotFound`

Closes #5129.

## Demo

These were the original redundant and misleading error messages produced
at the function definition side:


![image](https://github.com/FuelLabs/sway/assets/4142833/92ad3991-3f5b-40f4-b206-82287a650a70)

This is the new error message:


![image](https://github.com/FuelLabs/sway/assets/4142833/c17e64df-85c6-4797-aff3-743143468698)

In the case of an ambiguous trait, a detailed instructions are provided
on how to find and import the required trait:


![image](https://github.com/FuelLabs/sway/assets/4142833/cf852f66-29bf-4918-8878-ea02533c1164)

## Checklist

- [x] I have linked to any relevant issues.
- [x] I have commented my code, particularly in hard-to-understand
areas.
- [ ] I have updated the documentation where relevant (API docs, the
reference, and the Sway book).
- [x] I have added tests that prove my fix is effective or that my
feature works.
- [ ] I have added (or requested a maintainer to add) the necessary
`Breaking*` or `New Feature` labels where relevant.
- [x] I have done my best to ensure that my PR adheres to [the Fuel Labs
Code Review
Standards](https://github.com/FuelLabs/rfcs/blob/master/text/code-standards/external-contributors.md).
- [x] I have requested a review from the relevant team or maintainers.