Fixes error while using From trait that returns a tuple.
Instead of `Trait "MyFrom<u256>" is not implemented for type "U".`
We now get `Trait "MyFrom<u256>" is not implemented for type "(_, _, _, _)".`

The error is followed by:
`Cannot infer type for type parameter "_". Insufficient type information provided. Try annotating its type.`

While Rust compiler is able to inference the type in this situation
it is non trivial to do it properly hence we should probably keep asking
for the type to be annotated.