Changes `ConcurrentSlab` to no longer clone values
`ConcurrentSlab` now stores and returns an Arc, and on `get` it only clones an Arc which is much cheaper than cloning the whole values.

The type engine is still cloning every TypeInfo we should be able to also avoid that and get further improvments in speed.

With these changes core + std-lib compilation improved from around 3 secs to 1.5 secs.