Changes `ConcurrentSlab` to store values in map.
With these changes `ConcurrentSlab` now stores values in map instead of a vector.

This change is necessary to allow us to remove specific ids from the `ConcurrentSlab`.

Before this, removing an index would change the vector and invalidate existing type and decls ids.

This is required for `TypeEngine` and `DeclEngine` `clear_module` methods to work properly.