Revert TE `slab` to hold `TypeSourceInfo` and perform GC on DE `parents` and TE `id_map` (#5428)
## Description
The current GC implementation was not cleaning up `parents` in the decl
engine and also was missing the `id_map` in the type engine. This also
changes the `slab` in the type engine to hold a `TypeSourceInfo` instead
of having seperate maps.

The `type_info` in the `TypeSourceInfo` is now an `Arc`.