Provide option to the e2e test binary to specify build profile.
For tests that have their ABI tested, or rely on their compiled
hashes (for deployment), since that changes with build profile,
the tests are marked as unsupported for debug profile.
(better to test with and have expected results for release
than the debug profile).

Add testing release profile in CI since the default is now debug.

Two tests in `should_fail` that are expected to fail due to
overflows are disabled because there's a bug in our IR gen.
A new Issue will be opened for this.