Optimizes `TraitMap` `insert` to not use `TypeEngine`
`TraitMap` `insert` was adding values to the `TypeEngine` just to do a
unification check on them afterwards.

This increased `TypeInfo`s in type engine quadratically with the number of
trait already inserted. Because for each new trait inserted a new `TypeInfo::Custom`
 would be inserted for each existing trait.

This was fixed by doing the unification check directly on the type parameters
 instead of constructing a temporary `TypeInfo::Custom` with them.

For `should_pass/language/generic_trait_constraints` which uses lib-std the number of entries in the type engine changed from >194k to <93k.
Much less memory consumption would be expected in a script with lots of traits.