Fixes error while using From trait that returns a tuple. (#5470)
## Description

Instead of `Trait "MyFrom<u256>" is not implemented for type "U".` We
now get `Trait "MyFrom<u256>" is not implemented for type "(_, _, _,
_)".`

The error is followed by:
`Cannot infer type for type parameter "_". Insufficient type information
provided. Try annotating its type.`

While Rust compiler is able to inference the type in this situation it
is non trivial to do it properly hence for now we should probably keep
asking for the type to be annotated and create an issue for it.

## Checklist

- [ ] I have linked to any relevant issues.
- [ ] I have commented my code, particularly in hard-to-understand
areas.
- [ ] I have updated the documentation where relevant (API docs, the
reference, and the Sway book).
- [x] I have added tests that prove my fix is effective or that my
feature works.
- [x] I have added (or requested a maintainer to add) the necessary
`Breaking*` or `New Feature` labels where relevant.
- [x] I have done my best to ensure that my PR adheres to [the Fuel Labs
Code Review
Standards](https://github.com/FuelLabs/rfcs/blob/master/text/code-standards/external-contributors.md).
- [x] I have requested a review from the relevant team or maintainers.

Co-authored-by: Joshua Batty <joshpbatty@gmail.com>