Fixes issue of missing type annotations on explicit returns.
Explicit returns did not have any type annotation causing the method disambiguation to fail.

The solution was to add a new field to the `TypeCheckContext` called `function_type_annotation`.
This field is initialized only in the type_check of function declarations.
The field is used to set the type_annotation one while doing the type check of explicit return expressions.

With this change `TypeCheckUnification` was also simplified as the unification is now done with `function_type_annotation`.
The result was a few functions that are no longer required to be removed.

Closes #5518